requirements.txt
requests
python-dotenv
tqdm
beautifulsoup4